+++ src/models/InstabugConfig.ts
@@ -20,6 +20,11 @@ export interface InstabugConfig {
   codePushVersion?: string;
 
   /**
+   * An optional flag to override SDK screenshot security behavior.
+   */
+  ignoreAndroidSecureFlag?: boolean;
+
+  /**
    * An optional network interception mode, this determines whether network interception
    * is done in the JavaScript side or in the native Android and iOS SDK side.
    *