+++ src/modules/Instabug.ts
@@ -275,6 +275,7 @@ const initializeNativeInstabug = (config: InstabugConfig) => {
     shouldEnableNativeInterception &&
       config.networkInterceptionMode === NetworkInterceptionMode.native,
     config.codePushVersion,
+    config.ignoreAndroidSecureFlag,
   );
 };