+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/Instabug/Instabug-React-Native/compare/v15.1.0...dev)
+
+### Added
+
+- Add support ignore flag secure on android. ([#1394](https://github.com/Instabug/Instabug-React-Native/pull/1394))
+
 ## [15.0.1](https://github.com/Instabug/Instabug-React-Native/compare/v14.3.0...v15.0.1)
 
 ### Added